do code more portable

+++ bld/builder/c/sysposix.c
@@ -118,7 +118,7 @@ int SysRunCommand( const char *cmd )
 {
     int         my_std_output;
     int         my_std_error;
-    size_t      bytes_read;
+    ssize_t     bytes_read;
     int         rc;
     int         readpipe;
     char        buff[256 + 1];
@@ -138,7 +138,7 @@ int SysRunCommand( const char *cmd )
     }
     if( readpipe != -1 ) {
         while( (bytes_read = read( readpipe, buff, sizeof( buff ) - 1 )) != 0 ) {
-            if( (ssize_t)bytes_read == -1 )
+            if( bytes_read == -1 )
                 break;
             buff[bytes_read] = '\0';
             Log( Quiet, "%s", buff );
